Cartas SuperTrunfo tema2

+++ CartasSuperTrunfo.c
@@ -64,7 +64,7 @@ int main() {
 
     printf("O total de pontos turísticos é de %d\n\n", pontos);
 
-    densiPopulacao1 = (populacao / area); //calculando a densidade populacional.
+    densiPopulacao1 = ((double)populacao / area); //calculando a densidade populacional.
     pibcapita1 = (pib *1000000000.0) / populacao; //calculando o PIB per Capita.
 
     printf("Os dados foram salvos, obrigado por enviar.\nAgora vamos preencher os dados da proxima carta antes de concluir!\n\n");
@@ -112,7 +112,7 @@ int main() {
 
     printf("O total de pontos turísticos é de %d\n", pontos2);
 
-    densiPopulacao2 = (populacao2 / area2); //calculando a densidade populacional.
+    densiPopulacao2 = ((double)populacao2 / area2); //calculando a densidade populacional.
     pibcapita2 = (pib2 *1000000000.0) / populacao2; //calculando o PIB per Capita.
 
     printf("Perfeito!!!\n Estas são suas cartas, mantenha elas por perto ;D\n Até a próxima!\n\n");
